correct the reference to the actions as were moved to separate directory
